services/s18104_auto_label/pyproject.toml
[project]
name = "auto-label-service"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "requests>=2.28",
    "Pillow>=10.0",
    "numpy>=1.24",
    "opencv-python-headless>=4.8",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "supervision>=0.25.0",
    "av>=13.0",
]

[tool.setuptools.packages.find]
exclude = ["configs*", "tests*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
